feat: Add dryRun option

+++ main.go
@@ -17,9 +17,10 @@ var osGetwd = os.Getwd
 var osStat = os.Stat
 
 func main() {
-	arg := getArg()
+	target := getTarget()
+	dryRun := isDryRun()
 
-	request := makeRequest(arg)
+	request := makeRequest(target)
 	if request == nil {
 		return
 	}
@@ -35,15 +36,22 @@ func main() {
 	scanner := bufio.NewScanner(fp)
 
 	for scanner.Scan() {
-		delete(request.path, scanner.Text())
+		delete(request.path, scanner.Text(), dryRun)
 	}
 }
 
-func getArg() string {
+func getTarget() string {
 	if len(os.Args) < 2 {
 		return ""
 	}
 	return os.Args[1]
+}
+
+func isDryRun() bool {
+	if len(os.Args) < 3 {
+		return false
+	}
+	return strings.EqualFold(os.Args[2], "--dryRun=true")
 }
 
 func makeRequest(arg string) *Request {
@@ -84,14 +92,18 @@ func makeRequest(arg string) *Request {
 	}
 }
 
-func delete(path, pattern string) {
+func delete(path, pattern string, dryRun bool) {
 	files, err := filepath.Glob(path + pattern)
 	if err != nil {
 		log.Println("can not get: ", err)
 		return
 	}
 	for _, v := range files {
-		log.Println("remove: ", v)
-		os.Remove(v)
+		if dryRun {
+			log.Println("remove(without dryRun): ", v)
+		} else {
+			os.Remove(v)
+			log.Println("removed: ", v)
+		}
 	}
 }
